[REFACTOR] - 숫자 이동 로직 UseCase로 이동
