+++ app/src/main/java/com/jjsh/game2048/presentation/ui/main/MainViewModel.kt
@@ -1,12 +1,17 @@
 package com.jjsh.game2048.presentation.ui.main
 
 import androidx.lifecycle.ViewModel
+import com.jjsh.game2048.domain.usecase.MoveNumbersUseCase
 import com.jjsh.game2048.presentation.ui.view.MoveState
+import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
-import java.util.*
+import javax.inject.Inject
 
-class MainViewModel : ViewModel() {
+@HiltViewModel
+class MainViewModel @Inject constructor(
+    private val moveNumbersUseCase: MoveNumbersUseCase
+) : ViewModel() {
     private val _blockCount = MutableStateFlow(4)
     val blockCount: StateFlow<Int> get() = _blockCount
 
@@ -19,139 +24,12 @@ class MainViewModel : ViewModel() {
     val moveAction: (MoveState) -> Boolean = {
         beforeGameMap = gameMap.value.map { array -> array.copyOf() }
         when (it) {
-            MoveState.UP -> moveUpAction()
-            MoveState.DOWN -> moveDownAction()
-            MoveState.RIGHT -> moveRightAction()
-            MoveState.LEFT -> moveLeftAction()
+            MoveState.UP -> moveNumbersUseCase.moveUpAction(gameMap.value)
+            MoveState.DOWN -> moveNumbersUseCase.moveDownAction(gameMap.value)
+            MoveState.RIGHT -> moveNumbersUseCase.moveRightAction(gameMap.value)
+            MoveState.LEFT -> moveNumbersUseCase.moveLeftAction(gameMap.value)
         }
         compareMap()
-    }
-
-    private fun moveUpAction() {
-        val gameMap = this.gameMap.value
-        for (row in gameMap) {
-            val stack = Stack<Int>()
-
-            for (idx in row.indices) {
-                if (stack.isNotEmpty() && stack.peek() == row[idx]) {
-                    stack.add(stack.pop() + row[idx])
-                    stack.add(0)
-                } else if (row[idx] > 0) {
-                    stack.add(row[idx])
-                }
-            }
-
-            val newRow = mutableListOf<Int>()
-            while (stack.isNotEmpty()) {
-                val cur = stack.pop()
-                if (cur > 0) {
-                    newRow.add(0, cur)
-                }
-            }
-
-            var newRowIdx = 0
-            for (n in newRow) {
-                row[newRowIdx++] = n
-            }
-            for (i in newRowIdx..row.lastIndex) {
-                row[i] = 0
-            }
-        }
-    }
-
-    private fun moveDownAction() {
-        val gameMap = this.gameMap.value
-        for (row in gameMap) {
-            val stack = Stack<Int>()
-
-            for (idx in row.lastIndex downTo 0) {
-                if (stack.isNotEmpty() && stack.peek() == row[idx]) {
-                    stack.add(stack.pop() + row[idx])
-                    stack.add(0)
-                } else if (row[idx] > 0) {
-                    stack.add(row[idx])
-                }
-            }
-
-            val newRow = mutableListOf<Int>()
-            while (stack.isNotEmpty()) {
-                val cur = stack.pop()
-                if (cur > 0) {
-                    newRow.add(0, cur)
-                }
-            }
-
-            var newRowIdx = row.lastIndex
-            for (n in newRow) {
-                row[newRowIdx--] = n
-            }
-
-            for (i in newRowIdx downTo 0) {
-                row[i] = 0
-            }
-        }
-    }
-
-    private fun moveLeftAction() {
-        val gameMap = this.gameMap.value
-        for (c in gameMap[0].indices) {
-            val stack = Stack<Int>()
-            for (r in gameMap.indices) {
-                if (stack.isNotEmpty() && stack.peek() == gameMap[r][c]) {
-                    stack.add(stack.pop() + gameMap[r][c])
-                    stack.add(0)
-                } else if (gameMap[r][c] > 0) {
-                    stack.add(gameMap[r][c])
-                }
-            }
-            val newCol = mutableListOf<Int>()
-            while (stack.isNotEmpty()) {
-                val cur = stack.pop()
-                if (cur > 0) {
-                    newCol.add(0, cur)
-                }
-            }
-
-            var newColIdx = 0
-            for (n in newCol) {
-                gameMap[newColIdx++][c] = n
-            }
-
-            for (i in newColIdx..gameMap[c].lastIndex) {
-                gameMap[i][c] = 0
-            }
-        }
-    }
-
-    private fun moveRightAction() {
-        val gameMap = this.gameMap.value
-        for (c in gameMap[0].indices) {
-            val stack = Stack<Int>()
-            for (r in gameMap.lastIndex downTo 0) {
-                if (stack.isNotEmpty() && stack.peek() == gameMap[r][c]) {
-                    stack.add(stack.pop() + gameMap[r][c])
-                    stack.add(0)
-                } else if (gameMap[r][c] > 0) {
-                    stack.add(gameMap[r][c])
-                }
-            }
-            val newCol = mutableListOf<Int>()
-            while (stack.isNotEmpty()) {
-                val cur = stack.pop()
-                if (cur > 0) {
-                    newCol.add(0, cur)
-                }
-            }
-
-            var newColIdx = gameMap[c].lastIndex
-            for (n in newCol) {
-                gameMap[newColIdx--][c] = n
-            }
-
-            for (i in newColIdx downTo 0) {
-                gameMap[i][c] = 0
-            }
-        }
     }
 
     private fun compareMap(): Boolean {
